GPA convert

+++ springboot-offer-master/src/main/java/com/group7/entitiy/excel/GradeData.java
@@ -17,5 +17,5 @@ public class GradeData {
     private String grade;
 
     @ExcelProperty(index = 2)
-    private Float credits;
+    private Double credits;
 }
